引入 https://github.com/knezzz/ntp，解决APP端时间不是当前时间的问题

+++ lib/page/chat/chat_view.dart
@@ -426,7 +426,7 @@ class ChatPageState extends State<ChatPage> {
     if (msg.status != types.Status.sending) {
       return;
     }
-    int diff = DateTimeHelper.currentTimeMillis();
+    int diff = DateTimeHelper.currentTimeMillis() as int;
     if (diff > 1500) {
       // 检查为发送消息
       logic.sendWsMsg(logic.getMsgFromTmsg(
